code refactoring; non-blocking "workflow.run"
